+++ editor/widgets/leaf_editor.h
@@ -25,7 +25,7 @@ class LeafEditor : public Form {
 
 public:
 	LeafEditor(SharedResources *shared, Editor *editor, QWidget *parent);
-	void setFields(std::map<pg::Stem *, std::set<long>> instances);
+	void setFields(std::map<pg::Stem *, std::set<size_t>> instances);
 	bool addMaterial(ShaderParams params);
 	long removeMaterial(QString name);
 	void renameMaterial(QString before, QString after);
